requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
attrs==24.2.0
audioop-lts==0.2.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
idna==3.10
multidict==6.1.0
mutagen==1.47.0
propcache==0.2.0
py-cord==2.6.1
pycparser==2.22
pycryptodomex==3.21.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
websockets==13.1
yarl==1.18.0
yt-dlp==2024.10.16.232911.dev0
Flask==3.0.0
matplotlib==3.8.0